+++ src/components/Tip/ExamplePhotos.jsx
