+++ src/components/Tip/Manual.jsx
