+++ src/components/Tip/TipSection.jsx
