+++ src/pages/LibraryPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 // src/components 폴더에서 필요한 컴포넌트들을 임포트합니다.
-import HeaderBack from '../components/HeaderBack.jsx';
+import HeaderBack from '../components/Header/HeaderBack.jsx';
 import ImageGrid from '../components/ImageGrid/ImageGrid.jsx';
 import Navigation from '../components/Navigation/Navigation.jsx';
