+++ src/pages/RoomType.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import HeaderBack from '../components/HeaderBack';
+import HeaderBack from '../components/Header/HeaderBack';
 import ButtonAction from '../components/Button/ButtonAction';
 import ButtonS from '../components/Button/ButtonS';
 import Title from '../components/Title/Title';
